Modified Code according to template
